requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
cachetools==5.3.0
certifi==2022.12.7
cffi
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.1.0
cycler==0.11.0
filelock
fonttools==4.42.1
gmpy2
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.5.0
importlib-resources==6.0.1
Jinja2 
kiwisolver==1.4.5
lit==16.0.5
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.2
mkl-fft==1.3.6
mkl-random 
mkl-service==2.4.0
mpmath 
networkx
numpy
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
packaging==23.1
Pillow==8.4.0
progress==1.6
protobuf==4.22.3
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch 
torch-dct==0.1.6
torchaudio==0.10.0
torchvision==0.11.1
tqdm==4.66.1
transforms3d==0.4.1
triton==2.0.0
typing_extensions
ultralytics==8.0.199
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
